pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-ats-scorer"
version = "1.0.0"
description = "A multi-agent system for analyzing resumes against job descriptions and providing ATS compatibility scores"
authors = [
    {name = "Resume ATS Scorer Team", email = "support@resumeatsscorer.com"}
]
readme = "readme.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "crewai>=0.114.0",
    "pydantic>=2.11.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "python-multipart>=0.0.9",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.3",
    "spacy>=3.7.2",
    "gunicorn>=21.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.1.11",
    "langchain-openai>=0.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "coverage>=7.4.1",
    "black>=24.2.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
source = ["resume_ats_scorer"]
branch = true
omit = [
    "tests/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]

[project.urls]
"Homepage" = "https://github.com/deepakalevoor2/resume-ats-scorer"
"Bug Tracker" = "https://github.com/deepakalevoor2/resume-ats-scorer/issues"